fix: normalize dtype across audio/video/image pipeline to prevent float16 export failures
Root cause: models running in float16 precision produce float16 waveforms/frames.
soundfile (WAV) only accepts float32/float64/int16/int32; PIL/ImageIO/FFmpeg
require uint8. float16 data propagated through AudioData to export nodes,
causing 'dtype must be one of [float32, float64, int16, int32] and not float16'.

## Root cause fix: AudioData.__init__ (core/types.py)
- Auto-normalize waveform to float32 at construction time
- Checks for float16/other dtypes and converts to float32
- Eliminates the problem at the source for all downstream nodes

## Source node fixes (convert at generation time)
- music_generator.py: convert float16 numpy array to float32 after extraction
- sound_effect.py: explicit astype(np.float32) on AudioLDM2 output
- background_remover.py: .float() before ToPILImage to prevent float16 tensor issues

## Export node defense-in-depth (convert before writing)
- multi_format_exporter.py _export_audio: dtype check before sf.write
- multi_format_exporter.py _export_gif: uint8 conversion for numpy frames
- video_encoder.py _prepare_audio_input: dtype check before sf.write
- video_encoder.py _encode_frames: uint8 conversion for numpy frames before FFmpeg
- livestream.py: uint8 conversion for numpy frames before FFmpeg
- audio/_base.py _save_audio: dtype check before sf.write

## Video node fixes (convert before processing)
- video/_base.py _tensor_to_frames: float16→float32 after numpy conversion
- video/_base.py _save_video: uint8 conversion for numpy frames before imageio
- text_to_video.py: float16→float32 after numpy extraction
- video_continuation.py: float16→float32 after numpy extraction
- frame_extractor.py: uint8 conversion for numpy/path output formats

## Image node fixes
- image/_base.py _ensure_pil_image: uint8 conversion for float numpy arrays

All conversions use isinstance(arr, np.ndarray) guard to avoid issues with
MagicMock in test environments. Pattern follows existing best practice in
realtime_renderer.py (which already had this conversion).

All 953 tests pass (1 pre-existing skip, 0 regressions).

+++ mosaic/nodes/audio/sound_effect.py
@@ -174,14 +174,16 @@ class SoundEffectGenerator(BaseAudioNode):
                 import numpy as np  # type: ignore
 
                 waveform = audios[0]
+                # 显式转 float32，避免 float16 透传到导出环节
                 if isinstance(waveform, np.ndarray):
+                    waveform = waveform.astype(np.float32)
                     # 确保是 1D
                     if waveform.ndim > 1:
                         waveform = self._to_mono(waveform)
             else:
                 import numpy as np  # type: ignore
 
-                waveform = np.array([])
+                waveform = np.array([], dtype=np.float32)
 
             actual_duration = self._get_duration(waveform, sample_rate)
         except Exception as exc:
